+++ src/main/groovy/ru/vyarus/gradle/plugin/animalsniffer/AnimalSnifferPlugin.groovy
@@ -82,6 +82,7 @@ class AnimalSnifferPlugin implements Plugin<Project> {
     private void registerShortcuts() {
         ExtraPropertiesExtension props = project.extensions.extraProperties
         // to allow custom tasks declaration without package
+        props.set(AnimalSniffer.simpleName, AnimalSniffer)
         props.set(BuildSignatureTask.simpleName, BuildSignatureTask)
         props.set(SignatureInfoTask.simpleName, SignatureInfoTask)
     }