+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/ReportConfigurationKitTest.groovy
@@ -5,57 +5,9 @@ import org.gradle.testkit.runner.TaskOutcome
 
 /**
  * @author Vyacheslav Rusakov
- * @since 09.02.2023
+ * @since 18.11.2024
  */
-class LegacyKitTest extends AbstractKitTest {
-    public static final String GRADLE_VERSION = '5.1'
-
-    def "Check violation detection without cache task"() {
-        setup:
-        build """
-            plugins {
-                id 'java'
-                id 'ru.vyarus.animalsniffer'
-            }
-
-            animalsniffer {
-                ignoreFailures = true
-            }
-
-            repositories { mavenCentral()}
-            dependencies {
-                signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
-                implementation 'org.slf4j:slf4j-api:1.7.25'
-            }
-
-        """
-        fileFromClasspath('src/main/java/invalid/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java')
-//        debug()
-
-        when: "run task"
-        BuildResult result = runVer(GRADLE_VERSION, 'check', '--warning-mode', 'all')
-
-        then: "task successful"
-        result.task(':check').outcome == TaskOutcome.SUCCESS
-
-        then: "found 2 violations"
-        result.output.contains("2 AnimalSniffer violations were found in 1 files")
-        result.output.replaceAll('\r', '').contains(
-                """[Undefined reference] invalid.(Sample.java:11)
-  >> int Boolean.compare(boolean, boolean)
-
-[Undefined reference] invalid.(Sample.java:16)
-  >> java.nio.file.Path java.nio.file.Paths.get(String, String[])
-""")
-
-        then: "report correct"
-        File file = file('/build/reports/animalsniffer/main.text')
-        file.exists()
-        file.readLines() == [
-                "invalid.Sample:11  Undefined reference: int Boolean.compare(boolean, boolean)",
-                "invalid.Sample:16  Undefined reference: java.nio.file.Path java.nio.file.Paths.get(String, String[])"
-        ]
-    }
+class ReportConfigurationKitTest extends AbstractKitTest {
 
     def "Check file report path override"() {
         setup:
@@ -77,7 +29,6 @@ class LegacyKitTest extends AbstractKitTest {
             
             tasks.withType(AnimalSniffer) {
                 reports.text {
-                    enabled = true
                     destination = file('build/custom.txt')
                 }
             }