+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 * Update animalsniffer 1.23 -> 1.24
+* Fix gradle 8.11 support (#100)
+* Fix file report disabling
+* Add AnimalSniffer task shortcut for simpler reports configuration
 
 ### 1.7.1 (2023-07-05)
 * Update animalsniffer 1.22 -> 1.23